auth done

+++ src/menu/menu.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Menu } from 'src/typeorm';
 import { Repository } from 'typeorm';
@@ -10,51 +14,61 @@ export class MenuService {
     @InjectRepository(Menu)
     private readonly menuRepository: Repository<Menu>,
   ) {}
-  createMenu(createMenuDto: CreateMenuDto) {
-    // const prodId = Math.random().toString();
-    // const newMenu = this.menuRepository.create(createMenuDto);
-    const newMenu = this.menuRepository.create(createMenuDto);
-    console.log(
-      '🚀 ~ file: menu.service.ts:16 ~ MenuService ~ createMenu ~ newMenu:',
-      newMenu,
-    );
-    newMenu.restaurant = createMenuDto.restaurantId;
-    return this.menuRepository.save(newMenu);
-    // return prodId;
+  createMenu(createMenuDto: CreateMenuDto, req): string {
+    const { role } = req.user;
+    if (role == 'chef') {
+      const newMenu = this.menuRepository.create(createMenuDto);
+      newMenu.restaurant = createMenuDto.restaurantId;
+      this.menuRepository.save(newMenu);
+      return 'Item added';
+    }
+    throw new UnauthorizedException();
   }
 
-  getMenu() {
-    // return this.menuRepository.find();
+  getMenu(): Promise<Menu[]> {
     return this.menuRepository.find({ relations: ['restaurant'] });
   }
 
-  getSingleMenu(menuId: string) {
+  getSingleMenu(menuId: string): Promise<Menu> {
     return this.findMenu(menuId);
   }
 
-  async updateMenu(menuId: string, updateMenuDto: CreateMenuDto) {
-    const menu = await this.findMenu(menuId);
-    if (updateMenuDto.name) {
-      menu.name = updateMenuDto.name;
+  async updateMenu(
+    menuId: string,
+    updateMenuDto: CreateMenuDto,
+    req,
+  ): Promise<Menu> {
+    const { role } = req.user;
+    if (role == 'chef') {
+      const menu = await this.findMenu(menuId);
+      if (updateMenuDto.name) {
+        menu.name = updateMenuDto.name;
+      }
+      if (updateMenuDto.description) {
+        menu.description = updateMenuDto.description;
+      }
+      if (updateMenuDto.image) {
+        menu.image = updateMenuDto.image;
+      }
+      if (updateMenuDto.price) {
+        menu.price = updateMenuDto.price;
+      }
+      return this.menuRepository.save(menu);
     }
-    if (updateMenuDto.description) {
-      menu.description = updateMenuDto.description;
-    }
-    if (updateMenuDto.image) {
-      menu.image = updateMenuDto.image;
-    }
-    if (updateMenuDto.price) {
-      menu.price = updateMenuDto.price;
-    }
-    return this.menuRepository.save(menu);
+    throw new UnauthorizedException();
   }
 
-  async deleteMenu(menuId: string) {
-    const menu = await this.findMenu(menuId);
-    this.menuRepository.delete(menu);
+  async deleteMenu(menuId: string, req): Promise<string> {
+    const { role } = req.user;
+    if (role == 'chef') {
+      const menu = await this.findMenu(menuId);
+      this.menuRepository.delete(menu);
+      return 'Item deleted';
+    }
+    throw new UnauthorizedException();
   }
 
-  private async findMenu(id: string) {
+  private async findMenu(id: string): Promise<Menu> {
     const menu = await this.menuRepository.findOneBy({ id });
     if (!menu) {
       throw new NotFoundException('Could not find menu.');
